// frontend/src/services/api/models/Department.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type Department = {
    dept_code: number;
    dept_name: string;
    dept_parentcode: number;
    useCode?: (boolean | null);
    dept_operationmode?: (number | null);
    middleware_id?: (number | null);
    defaultDepartment?: (number | null);
    company_id: number;
    lineToken?: (string | null);
    description?: (string | null);
    id: number;
};